backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
mysql-connector-python==8.1.0
python-dateutil==2.8.2
regex==2023.8.8
requests==2.31.0
openai>=1.40.0
python-multipart==0.0.6
Pillow==10.0.0
PyPDF2==3.0.1
pytesseract==0.3.10
werkzeug==2.3.7
python-dotenv==1.0.0
# LangFlow and AI workflow dependencies
langflow>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.0
# Opik telemetry and tracing
opik>=0.1.0
# Additional workflow dependencies
pydantic>=2.0.0
httpx>=0.25.0